Fixed tests failure on old driver.
